Added support for Webdriver Standalone
The before function is currently already being called with the arguments of config, capabilities, browser.
Since the browser object will be the same, this change allows the service to be compatible with the CLI Testrunner
and WebDriver standalone. Documentation in the README and a new test has been added also.