test(request-order): add tests for the order of retrieved requests
The behavior in 4.1.10 and earlier versions is that requests are ordered by time of fulfillment (`orderBy: 'END'`). This should remain the default behavior, but users should be able to order by the time the request started as well (which is possibly more under their control, in terms of testing, too).
When pending requests are requested, this default sort will place them last (as they are not yet "completed").